requirements.txt
python-telegram-bot[job-queue]
aiohttp
python-dotenv
